Investigate Xlib errors when generating UI (#9)
* Removed AntTweakBar, included Nuklear, some changes to OpenGL setup

* Added first batch of UI controls

* Finished UI layout

* Finished UI

* Minor fixes

Removed commented out function from previous UI library
Changed exit on main
Use variable for viewport in main method

+++ src/Swarm.cpp
@@ -5,12 +5,13 @@
  * @author Fernando Ferreira
  */
 
+#include <random>
+#include <vector>
+
 #include <Agent.h>
 #include <Attractor.h>
-#include <random>
 #include <Swarm.h>
 #include <Triplet.h>
-#include <vector>
 
 const int FREEFORM       = 0;
 const int AVERAGE        = 1;
@@ -163,7 +164,7 @@ void Swarm::resetAttractors()
 }
 
 
-void Swarm::addAttractor(long pitch, int tone = -1)
+void Swarm::addAttractor(int pitch, int tone = -1)
 {
     Attractor attractor(pitch, tone);
     attractors.push_back(attractor);
